Pass by reference
Kla::createuser(): explicitly pass a reference to the @members array to
the Kla::addmembers() sub. Since this is perl, and OO, prototypes don't
work here, resulting in perl using it /without/ the reference, which in
turn results in us getting a scalar rather than an array which we
expect. Bummer.